+++ src/starkbank/tax_payment.clj
@@ -1,38 +1,38 @@
-(ns starkbank.brcode-payment
-    "When you initialize a BrcodePayment, the entity will not be automatically
-    created in the Stark Bank API. The 'create' function sends the objects
-    to the Stark Bank API and returns the list of created objects.
-  
-    ## Parameters (required):
-      - `:brcode` [string]: String loaded directly from the QRCode or copied from the invoice. ex: \"00020126580014br.gov.bcb.pix0136a629532e-7693-4846-852d-1bbff817b5a8520400005303986540510.005802BR5908T'Challa6009Sao Paulo62090505123456304B14A\"
-      - `:tax-id` [string]: receiver tax ID (CPF or CNPJ) with or without formatting. ex: \"01234567890\" or \"20.018.183/0001-80\"
-      - `:description` [string]: Text to be displayed in your statement (min. 10 characters). ex: \"payment ABC\"
-  
-    ## Parameters (optional):
-      - `:amount` [integer, default nil]: amount automatically calculated from line or barCode. ex: 23456 (= R$ 234.56)
-      - `:scheduled` [string, default now]: payment scheduled date or datetime. ex: \"2020-11-25T17:59:26.249976+00:00\"
-      - `:tags` [list of strings, default nil]: list of strings for tagging
-  
-    ## Attributes (return-only):
-      - `:id` [string, default nil]: unique id returned when payment is created. ex: \"5656565656565656\"
-      - `:status` [string, default nil]: current payment status. ex: \"success\" or \"failed\"
-      - `:type` [string, default nil]: brcode type. ex: \"static\" or \"dynamic\"
-      - `:fee` [integer, default nil]: fee charged when the brcode payment is created. ex: 200 (= R$ 2.00)
-      - `:transaction-ids` [list of strings, default nil]: ledger transaction ids linked to this boleto. ex: [\"19827356981273\"] 
-      - `:updated` [string, default nil]: latest update datetime for the payment. ex: \"2020-11-25T17:59:26.249976+00:00\"
-      - `:created` [string, default nil]: creation datetime for the payment. ex: \"2020-11-25T17:59:26.249976+00:00\""
-    (:refer-clojure :exclude [get set update])
-    (:import [com.starkbank BrcodePayment])
-    (:use [starkbank.user]
-          [clojure.walk]))
-  
+(ns starkbank.tax-payment
+  "When you initialize a TaxPayment, the entity will not be automatically
+  created in the Stark Bank API. The 'create' function sends the objects
+  to the Stark Bank API and returns the list of created objects.
+
+  ## Parameters (conditionally required):
+    - `:line` [string, default nil]: Number sequence that describes the payment. Either 'line' or 'bar-code' parameters are required. If both are sent, they must match. ex: \"34191.09008 63571.277308 71444.640008 5 81960000000062\"
+    - `:bar-code` [string, default nil]: Bar code number that describes the payment. Either 'line' or 'barCode' parameters are required. If both are sent, they must match. ex: \"34195819600000000621090063571277307144464000\"
+
+  ## Parameters (required):
+    - `:description` [string]: Text to be displayed in your statement (min. 10 characters). ex: \"payment ABC\"
+
+  ## Parameters (optional):
+    - `:scheduled` [string, default today]: payment scheduled date. ex: ~D[2020-03-25]
+    - `:tags` [list of strings]: list of strings for tagging
+
+  Attributes (return-only):
+    - `:id` [string, default nil]: unique id returned when the payment is created. ex: \"5656565656565656\"
+    - `:type` [string, default nil]: tax type. ex: \"das\"
+    - `:status` [string, default nil]: current payment status. ex: \"processing\" or \"success\"
+    - `:amount` [integer, default nil]: amount automatically calculated from line or bar-code. ex: 23456 (= R$ 234.56)
+    - `:fee` [integer, default nil]: fee charged when a utility payment is created. ex: 200 (= R$ 2.00)
+    - `:updated` [string, default nil]: latest update datetime for the payment. ex: \"2020-03-26T19:32:35.418698+00:00\"
+    - `:created` [string, default nil]: creation datetime for the payment. ex: \"2020-03-26T19:32:35.418698+00:00\""
+  (:refer-clojure :exclude [get set])
+  (:import [com.starkbank TaxPayment])
+  (:use [starkbank.user]
+        [clojure.walk]))
+
 (defn- clojure-to-java
   ([clojure-map]
     (let [{
-      brcode "brcode"
-      tax-id "tax-id"
       description "description"
-      amount "amount"
+      line "line"
+      bar-code "bar-code"
       scheduled "scheduled"
       tags "tags"
     }
@@ -40,12 +40,11 @@
 
       (defn- apply-java-hashmap [x] (java.util.HashMap. x))
       
-      (BrcodePayment. (java.util.HashMap.
+      (TaxPayment. (java.util.HashMap.
         {
-          "brcode" brcode
-          "taxId" tax-id
           "description" description
-          "amount" (if (nil? amount) nil (Long. amount))
+          "line" line
+          "barCode" bar-code
           "scheduled" scheduled
           "tags" (if (nil? tags) nil (into-array String tags))
         }
@@ -55,18 +54,17 @@
   ([java-object]
     {
       :id (.id java-object)
-      :brcode (.brcode java-object)
-      :tax-id (.taxId java-object)
+      :line (.line java-object)
+      :bar-code (.barCode java-object)
       :description (.description java-object)
-      :amount (.amount java-object)
-      :scheduled (.scheduled java-object)
-      :tags (.tags java-object)
+      :tags (into [] (.tags java-object))
+      :scheduled  (.scheduled java-object)
       :status (.status java-object)
-      :type (.type java-object)
+      :amount (.amount java-object)
       :fee (.fee java-object)
-      :transaction-ids (into [] (.transactionIds java-object))
-      :updated (.updated java-object)
-      :created (.created java-object)
+      :type (.type java-object)
+      :updated  (.updated java-object)
+      :created  (.created java-object)
     }))
 
 (defn- clojure-query-to-java
@@ -112,40 +110,30 @@
           "status" status
         }
       ))))
-      
-(defn- clojure-update-to-java
-  ([clojure-map]
-    (let [{
-      status "status"
-    } (stringify-keys clojure-map)]
-      (java.util.HashMap.
-      {
-        "status" status
-      }))))
 
 (defn create
-  "Send a list of BrcodePayment maps for creation in the Stark Bank API
+  "Send a list of TaxPayment objects for creation in the Stark Bank API
 
   ## Parameters (required):
-    - `payments` [list of BrcodePayment maps]: list of BrcodePayment maps to be created in the API
+    - `payments` [list of TaxPayment maps]: list of TaxPayment maps to be created in the API
 
   ## Options:
     - `:user` [Project or Organization, default nil]: Project or Organization map returned from starkbank.user/project or starkbank.user/organization. Only necessary if starkbank.settings/user has not been set.
 
   ## Return:
-    - list of BrcodePayment maps with updated attributes"
+    - list of TaxPayment maps with updated attributes"
   ([payments]
     (def java-payments (map clojure-to-java payments))
-    (def created-java-payments (BrcodePayment/create java-payments))
+    (def created-java-payments (TaxPayment/create java-payments))
     (map java-to-clojure created-java-payments))
 
   ([payments, user]
     (def java-payments (map clojure-to-java payments))
-    (def created-java-payments (BrcodePayment/create java-payments (#'starkbank.user/get-java-user user)))
+    (def created-java-payments (TaxPayment/create java-payments (#'starkbank.user/get-java-user user)))
     (map java-to-clojure created-java-payments)))
 
 (defn query
-  "Receive a stream of BrcodePayment maps previously created in the Stark Bank API.
+  "Receive a stream of TaxPayment objects previously created in the Stark Bank API.
   Use this function instead of page if you want to stream the objects without worrying about cursors and pagination.
 
   ## Options:
@@ -153,25 +141,25 @@
     - `:after` [string, default nil]: date filter for maps created only after specified date. ex: \"2020-3-10\"
     - `:before` [string, default nil]: date filter for maps created only before specified date. ex: \"2020-3-10\"
     - `:tags` [list of strings, default nil]: tags to filter retrieved maps. ex: [\"tony\", \"stark\"]
-    - `:ids` [list of strings, default nil]: list of ids to filter retrieved objects. ex: [\"5656565656565656\", \"4545454545454545\"]
+    - `:ids` [list of strings, default nil]: list of ids to filter retrieved maps. ex: [\"5656565656565656\", \"4545454545454545\"]
     - `:status` [string, default nil]: filter for status of retrieved maps. ex: \"success\"
     - `:user` [Project or Organization, default nil]: Project or Organization map returned from starkbank.user/project or starkbank.user/organization. Only necessary if starkbank.settings/user has not been set.
 
   ## Return:
-    - stream of BrcodePayment maps with updated attributes"
+    - stream of UtilityPayment maps with updated attributes"
   ([]
-    (map java-to-clojure (BrcodePayment/query)))
+    (map java-to-clojure (TaxPayment/query)))
 
   ([params]
     (def java-params (clojure-query-to-java params))
-    (map java-to-clojure (BrcodePayment/query java-params)))
+    (map java-to-clojure (TaxPayment/query java-params)))
 
   ([params, user] 
     (def java-params (clojure-query-to-java params))
-    (map java-to-clojure (BrcodePayment/query java-params (#'starkbank.user/get-java-user user)))))
+    (map java-to-clojure (TaxPayment/query java-params (#'starkbank.user/get-java-user user)))))
 
 (defn page
-  "Receive a list of up to 100 BrcodePayment maps previously created in the Stark Bank API and the cursor to the next page.
+  "Receive a list of up to 100 TaxPayment objects previously created in the Stark Bank API and the cursor to the next page.
   Use this function instead of query if you want to manually page your requests.
 
   ## Options:
@@ -180,7 +168,7 @@
     - `:after` [string, default nil]: date filter for maps created only after specified date. ex: \"2020-3-10\"
     - `:before` [string, default nil]: date filter for maps created only before specified date. ex: \"2020-3-10\"
     - `:tags` [list of strings, default nil]: tags to filter retrieved maps. ex: [\"tony\", \"stark\"]
-    - `:ids` [list of strings, default nil]: list of ids to filter retrieved objects. ex: [\"5656565656565656\", \"4545454545454545\"]
+    - `:ids` [list of strings, default nil]: list of ids to filter retrieved maps. ex: [\"5656565656565656\", \"4545454545454545\"]
     - `:status` [string, default nil]: filter for status of retrieved maps. ex: \"success\"
     - `:user` [Project or Organization, default nil]: Project or Organization map returned from starkbank.user/project or starkbank.user/organization. Only necessary if starkbank.settings/user has not been set.
 
@@ -189,23 +177,23 @@
       - `:payments`: list of payment maps with updated attributes
       - `:cursor`: cursor string to retrieve the next page of payments"
   ([]
-    (def payment-page (BrcodePayment/page))
+    (def payment-page (TaxPayment/page))
     (def cursor (.cursor payment-page))
     (def payments (map java-to-clojure (.payments payment-page)))
     {:payments payments, :cursor cursor})
 
   ([params]
     (def java-params (clojure-page-to-java params))
-    (def payment-page (BrcodePayment/page java-params))
+    (def payment-page (TaxPayment/page java-params))
     {:payments (map java-to-clojure (.payments payment-page)), :cursor (.cursor payment-page)})
 
   ([params, user] 
     (def java-params (clojure-page-to-java params))
-    (def payment-page (BrcodePayment/page java-params (#'starkbank.user/get-java-user user)))
+    (def payment-page (TaxPayment/page java-params (#'starkbank.user/get-java-user user)))
     {:payments (map java-to-clojure (.payments payment-page)), :cursor (.cursor payment-page)}))
-    
+
 (defn get
-  "Receive a single BrcodePayment map previously created by the Stark Bank API by passing its id
+  "Receive a single TaxPayment object previously created by the Stark Bank API by passing its id
 
   ## Parameters (required):
     - `:id` [string]: map unique id. ex: \"5656565656565656\"
@@ -214,20 +202,41 @@
     - `:user` [Project or Organization, default nil]: Project or Organization map returned from starkbank.user/project or starkbank.user/organization. Only necessary if starkbank.settings/user has not been set.
 
   ## Return:
-    - BrcodePayment map with updated attributes"
+    - TaxPayment map with updated attributes"
   ([id]
     (java-to-clojure
-      (BrcodePayment/get id)))
+      (TaxPayment/get id)))
 
   ([id, user]
     (java-to-clojure
-      (BrcodePayment/get
+      (TaxPayment/get
+        id
+        (#'starkbank.user/get-java-user user)))))
+
+(defn delete
+  "Delete a TaxPayment entity previously created in the Stark Bank API
+
+  ## Parameters (required):
+    - `:id` [string]: UtilityPayment unique id. ex: \"5656565656565656\"
+
+  ## Options:
+    - `:user` [Project or Organization, default nil]: Project or Organization map returned from starkbank.user/project or starkbank.user/organization. Only necessary if starkbank.settings/user has not been set.
+
+  ## Return:
+    - deleted TaxPayment map"
+  ([id]
+    (java-to-clojure
+      (TaxPayment/delete id)))
+
+  ([id, user]
+    (java-to-clojure
+      (TaxPayment/delete
         id
         (#'starkbank.user/get-java-user user)))))
 
 (defn pdf
-  "Receive a single BrcodePayment pdf file generated in the Stark Bank API by passing its id.
-  Only valid for brcode payments with \"success\" status.
+  "Receive a single TaxPayment pdf file generated in the Stark Bank API by passing its id.
+  Only valid for tax payments with \"success\" status.
 
   ## Parameters (required):
     - `:id` [string]: map unique id. ex: \"5656565656565656\"
@@ -236,54 +245,32 @@
     - `:user` [Project or Organization, default nil]: Project or Organization map returned from starkbank.user/project or starkbank.user/organization. Only necessary if starkbank.settings/user has not been set.
 
   ## Return:
-    - BrcodePayment pdf file content"
+    - TaxPayment pdf file content"
   ([id]
     (clojure.java.io/input-stream
-      (BrcodePayment/pdf id)))
+      (TaxPayment/pdf id)))
 
   ([id, user]
     (clojure.java.io/input-stream
-      (BrcodePayment/pdf
+      (TaxPayment/pdf
         id
         (#'starkbank.user/get-java-user user)))))
-  
-(defn update
-  "Update a BrcodePayment by passing id.
 
-  ## Parameters (required):
-    - `:id` [list of strings]: BrcodePayment unique ids. ex: \"5656565656565656\"
 
-  ## Parameters (optional):
-    - `:status` [string]: If the BrcodePayment hasn't been paid yet, you may cancel it by passing \"canceled\" in the status
-
-  ## Return:
-    - target BrcodePayment with updated attributes"
-  ([id, params]
-    (java-to-clojure
-    (BrcodePayment/update id (clojure-update-to-java params))))
-
-  ([id, params, user]
-    (java-to-clojure
-    (BrcodePayment/update
-      id
-      (clojure-update-to-java params)
-      (#'starkbank.user/get-java-user user)))))
-
-(ns starkbank.brcode-payment.log
-  "Every time a BrcodePayment entity is modified, a corresponding BrcodePayment.Log
-  is generated for the entity. This log is never generated by the
-  user, but it can be retrieved to check additional information
-  on the BrcodePayment.
+(ns starkbank.tax-payment.log
+  "Every time a TaxPayment entity is modified, a corresponding TaxPayment.Log
+  is generated for the entity. This log is never generated by the user, but it can
+  be retrieved to check additional information on the TaxPayment.
 
   ## Attributes:
     - `:id` [string]: unique id returned when the log is created. ex: \"5656565656565656\"
-    - `:payment` [BrcodePayment]: BrcodePayment entity to which the log refers to.
-    - `:errors` [list of strings]: list of errors linked to this BrcodePayment event.
-    - `:type` [string]: type of the BrcodePayment event which triggered the log creation. ex: \"processing\" or \"success\"
+    - `:payment` [UtilityPayment]: UtilityPayment entity to which the log refers to.
+    - `:errors` [list of strings]: list of errors linked to this UtilityPayment event.
+    - `:type` [string]: type of the UtilityPayment event which triggered the log creation. ex: \"processing\" or \"success\"
     - `:created` [string]: creation datetime for the log. ex: \"2020-03-26T19:32:35.418698+00:00\""
   (:refer-clojure :exclude [get set])
-  (:import [com.starkbank BrcodePayment$Log])
-  (:require [starkbank.brcode-payment :as payment])
+  (:import [com.starkbank TaxPayment$Log])
+  (:require [starkbank.tax-payment :as payment])
   (:use [starkbank.user]
         [clojure.walk]))
 
@@ -350,11 +337,11 @@
     - Log map with updated attributes"
   ([id]
     (java-to-clojure
-      (BrcodePayment$Log/get id)))
+      (TaxPayment$Log/get id)))
 
   ([id, user]
     (java-to-clojure
-      (BrcodePayment$Log/get
+      (TaxPayment$Log/get
         id
         (#'starkbank.user/get-java-user user)))))
 
@@ -367,24 +354,24 @@
     - `:after` [string, default nil]: date filter for entities created only after specified date. ex: \"2020-3-10\"
     - `:before` [string, default nil]: date filter for entities created only before specified date. ex: \"2020-3-10\"
     - `:types` [list of strings, default nil]: filter retrieved entities by event types. ex: \"processing\" or \"success\"
-    - `:payment-ids` [list of strings, default nil]: list of BrcodePayment ids to filter retrieved entities. ex: [\"5656565656565656\", \"4545454545454545\"]
+    - `:payment-ids` [list of strings, default nil]: list of UtilityPayment ids to filter retrieved entities. ex: [\"5656565656565656\", \"4545454545454545\"]
     - `:user` [Project or Organization, default nil]: Project or Organization map returned from starkbank.user/project or starkbank.user/organization. Only necessary if starkbank.settings/user has not been set.
 
   ## Return:
     - stream of Log maps with updated attributes"
   ([]
-    (map java-to-clojure (BrcodePayment$Log/query)))
+    (map java-to-clojure (TaxPayment$Log/query)))
 
   ([params]
     (def java-params (clojure-query-to-java params))
-    (map java-to-clojure (BrcodePayment$Log/query java-params)))
+    (map java-to-clojure (TaxPayment$Log/query java-params)))
 
   ([params, user] 
     (def java-params (clojure-query-to-java params))
-    (map java-to-clojure (BrcodePayment$Log/query java-params (#'starkbank.user/get-java-user user)))))
-    
+    (map java-to-clojure (TaxPayment$Log/query java-params (#'starkbank.user/get-java-user user)))))
+
 (defn page
-  "Receive a list of up to 100 BrcodePayment.Log maps previously created in the Stark Bank API and the cursor to the next page.
+  "Receive a list of up to 100 TaxPayment.Log maps previously created in the Stark Bank API and the cursor to the next page.
   Use this function instead of query if you want to manually page your requests.
 
   ## Options:
@@ -393,7 +380,7 @@
     - `:after` [string, default nil]: date filter for entities created only after specified date. ex: \"2020-3-10\"
     - `:before` [string, default nil]: date filter for entities created only before specified date. ex: \"2020-3-10\"
     - `:types` [list of strings, default nil]: filter retrieved entities by event types. ex: \"processing\" or \"success\"
-    - `:payment-ids` [list of strings, default nil]: list of BrcodePayment ids to filter retrieved entities. ex: [\"5656565656565656\", \"4545454545454545\"]
+    - `:payment-ids` [list of strings, default nil]: list of TaxPayment ids to filter retrieved entities. ex: [\"5656565656565656\", \"4545454545454545\"]
     - `:user` [Project or Organization, default nil]: Project or Organization map returned from starkbank.user/project or starkbank.user/organization. Only necessary if starkbank.settings/user has not been set.
 
   ## Return:
@@ -401,17 +388,17 @@
       - `:logs`: list of log maps with updated attributes
       - `:cursor`: cursor string to retrieve the next page of logs"
   ([]
-    (def log-page (BrcodePayment$Log/page))
+    (def log-page (TaxPayment$Log/page))
     (def cursor (.cursor log-page))
     (def logs (map java-to-clojure (.logs log-page)))
     {:logs logs, :cursor cursor})
 
   ([params]
     (def java-params (clojure-page-to-java params))
-    (def log-page (BrcodePayment$Log/page java-params))
+    (def log-page (TaxPayment$Log/page java-params))
     {:logs (map java-to-clojure (.logs log-page)), :cursor (.cursor log-page)})
 
   ([params, user] 
     (def java-params (clojure-page-to-java params))
-    (def log-page (BrcodePayment$Log/page java-params (#'starkbank.user/get-java-user user)))
+    (def log-page (TaxPayment$Log/page java-params (#'starkbank.user/get-java-user user)))
     {:logs (map java-to-clojure (.logs log-page)), :cursor (.cursor log-page)}))