+++ src/starkbank/workspace.clj
@@ -174,7 +174,7 @@
     - `:id` [string]: map unique id. ex: \"5656565656565656\"
 
   ## Options:
-    - `:user` [Project or Organization]: Project or Organization map returned from starkbank.user/project or starkbank.user/organization. Only necessary if starkbank.settings/user has not been set.
+    - `:user` [Project or Organization, default nil]: Project or Organization map returned from starkbank.user/project or starkbank.user/organization. Only necessary if starkbank.settings/user has not been set.
 
   ## Return:
     - Workspace map with updated attributes"